+++ lib/src/wire/builders/change_notifier_builder.dart
@@ -62,18 +62,22 @@ class TextEditingControllerBuilder extends ChangeNotifierBuilder {
   TextEditingControllerBuilder(Application application) : super(application);
 
   @override
-  material.TextEditingController build(Json json, material.BuildContext? context) {
+  material.TextEditingController build(
+      Json json, material.BuildContext? context) {
     final id = json.data["id"];
 
     return application
         .make<StateManager>(WireDefinition.stateManager)
-        .get<TextEditingController>(WireDefinition.stateCallableRegistryState, id,
-            defaultValue: TextEditingController(material.TextEditingController()))
+        .get<TextEditingController>(
+            WireDefinition.stateCallableRegistryState, id,
+            defaultValue:
+                TextEditingController(material.TextEditingController()))
         .instance;
   }
 }
 
-class TextEditingController extends CallableRegistry<material.TextEditingController> {
+class TextEditingController
+    extends CallableRegistry<material.TextEditingController> {
   static const String methodClear = "clear";
   static const String methodDispose = "dispose";