+++ lib/src/wire/builders/validator_builder.dart
@@ -51,9 +51,11 @@ class RequiredValidatorBuilder extends PredefinedValidatorBuilder {
         return value ? null : json.data["errorMessage"];
       }
 
-      return (value == null || value.isEmpty)
-          ? json.data["errorMessage"]
-          : null;
+      if (value == null || (value is String && value.isEmpty)) {
+        return json.data["errorMessage"];
+      }
+
+      return null;
     };
   }
 }
@@ -93,7 +95,12 @@ class MinLengthValidatorBuilder extends PredefinedValidatorBuilder {
 
   @override
   FormFieldValidator build(Json json, material.BuildContext? context) {
+    final withoutSpaces = json.data["withoutSpaces"] ?? false;
+
     return (dynamic value) {
+      if (withoutSpaces) {
+        value = value.replaceAll(" ", "");
+      }
       return (value.length < json.data["min"])
           ? json.data["errorMessage"]
           : null;
@@ -106,7 +113,12 @@ class MaxLengthValidatorBuilder extends PredefinedValidatorBuilder {
 
   @override
   FormFieldValidator build(Json json, material.BuildContext? context) {
+    final withoutSpaces = json.data["withoutSpaces"] ?? false;
+
     return (dynamic value) {
+      if (withoutSpaces) {
+        value = value.replaceAll(" ", "");
+      }
       return (value.length > json.data["max"])
           ? json.data["errorMessage"]
           : null;
@@ -128,6 +140,22 @@ class RangeValidatorBuilder extends PredefinedValidatorBuilder {
         return json.data["errorMessage"];
       }
       return null;
+    };
+  }
+}
+
+class RegexValidatorBuilder extends PredefinedValidatorBuilder {
+  RegexValidatorBuilder(Application application) : super(application);
+
+  @override
+  FormFieldValidator build(Json json, material.BuildContext? context) {
+    final pattern = json.data["pattern"];
+    final caseSensitive = json.data["caseSensitive"] ?? true;
+
+    return (dynamic value) {
+      return hasMatch(pattern, value, caseSensitive: caseSensitive)
+          ? null
+          : json.data["errorMessage"];
     };
   }
 }