+++ lib/src/wire/definitions/json.dart
@@ -24,6 +24,7 @@ class JsonDefinition {
   static const form = 'Form';
   static const textFormField = 'TextFormField';
   static const checkboxFormField = 'CheckboxFormField';
+  static const radioGroupFormField = 'RadioGroupFormField';
   static const circularProgressIndicator = 'CircularProgressIndicator';
   static const statelessWidget = 'StatelessWidget';
   static const materialApp = 'MaterialApp';
@@ -32,6 +33,7 @@ class JsonDefinition {
   static const bottomAppBar = 'BottomAppBar';
   static const materialPage = 'MaterialPage';
   static const alertAction = 'AlertAction';
+  static const loadingAction = 'LoadingAction';
   static const navigationAction = 'NavigationAction';
   static const networkRequestAction = 'NetworkRequestAction';
   static const eventDispatchAction = 'EventDispatchAction';
@@ -41,6 +43,8 @@ class JsonDefinition {
   static const modalBottomSheetAction = 'ModalBottomSheetAction';
   static const updateStateAction = 'UpdateStateAction';
   static const updateNotifierAction = 'UpdateNotifierAction';
+  static const submitAction = 'SubmitAction';
+  static const validateAndSaveFormAction = 'ValidateAndSaveFormAction';
   static const scrollController = 'ScrollController';
   static const intrinsicHeight = 'IntrinsicHeight';
   static const floatingActionButton = 'FloatingActionButton';
@@ -58,6 +62,7 @@ class JsonDefinition {
   static const radioListTile = 'RadioListTile';
   static const listTile = 'ListTile';
   static const pingFormField = 'PingFormField';
+  static const overlay = 'Overlay';
   static const inputDecorationBuilder = 'InputDecoration';
   static const composeValidator = 'ComposeValidator';
   static const requiredValidator = 'RequiredValidator';
@@ -66,4 +71,5 @@ class JsonDefinition {
   static const maxLengthValidator = 'MaxLengthValidator';
   static const lengthValidator = 'LengthValidator';
   static const rangeValidator = 'RangeValidator';
+  static const regexValidator = 'RegexValidator';
 }