+++ lib/src/wire/provider/json_builder_provider.dart
@@ -50,6 +50,8 @@ class JsonBuilderProvider extends Provider {
         JsonDefinition.textFormField, () => TextFormFieldBuilder(app));
     app.singleton(
         JsonDefinition.checkboxFormField, () => CheckboxFormFieldBuilder(app));
+    app.singleton(
+        JsonDefinition.radioGroupFormField, () => RadioGroupFormFieldBuilder(app));
     app.singleton(JsonDefinition.circularProgressIndicator,
         () => CircularProgressIndicatorBuilder(app));
     app.singleton(
@@ -72,6 +74,8 @@ class JsonBuilderProvider extends Provider {
     app.singleton(JsonDefinition.listTile, () => ListTileBuilder(app));
     app.singleton(
         JsonDefinition.pingFormField, () => PingFormFieldBuilder(app));
+    app.singleton(
+        JsonDefinition.overlay, () => OverlayBuilder(app));
     //paintings builders
     app.singleton(JsonDefinition.inputDecorationBuilder,
         () => InputDecorationBuilder(app));
@@ -88,6 +92,7 @@ class JsonBuilderProvider extends Provider {
 
     //TODO make element_executor_provider
     app.singleton(JsonDefinition.alertAction, () => AlertActionExecutor(app));
+    app.singleton(JsonDefinition.loadingAction, () => LoadingActionExecutor(app));
 
     // dynamic value builders
 
@@ -130,6 +135,12 @@ class JsonBuilderProvider extends Provider {
     app.singleton(JsonDefinition.updateNotifierAction,
         () => UpdateNotifierActionExecutor(app));
 
+    app.singleton(JsonDefinition.submitAction,
+        () => SubmitActionExecutor(app));
+
+    app.singleton(JsonDefinition.validateAndSaveFormAction,
+        () => ValidateAndSaveFormActionExecutor(app));
+
     // form field validators
     app.singleton(
         JsonDefinition.composeValidator, () => ComposerValidatorBuilder(app));
@@ -145,5 +156,7 @@ class JsonBuilderProvider extends Provider {
         JsonDefinition.lengthValidator, () => LengthValidatorBuilder(app));
     app.singleton(
         JsonDefinition.rangeValidator, () => RangeValidatorBuilder(app));
+    app.singleton(
+        JsonDefinition.regexValidator, () => RegexValidatorBuilder(app));
   }
 }