+++ lib/src/wire/resources/paintings/border_side.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart' as material;
+import 'package:flutter_ping_wire/src/wire/resources/ui/border_style.dart';
 import 'painting.dart';
 import '../ui/color.dart';
 
@@ -11,10 +12,12 @@ abstract class AbstractBorderSide implements Painting {
 class BorderSide implements AbstractBorderSide {
   AbstractColor? color;
   double width;
+  BorderStyle? style;
 
   BorderSide({
     required this.color,
     required this.width,
+    this.style,
   });
 
   factory BorderSide.fromRawJson(String str) =>
@@ -27,6 +30,7 @@ class BorderSide implements AbstractBorderSide {
           : Color.findColor(json["color"]),
       width:
           json["width"] == null ? 1.0 : double.parse(json["width"].toString()),
+      style: json["style"] == null ? null : BorderStyle.fromJson(json["style"]),
     );
   }
 
@@ -39,6 +43,7 @@ class BorderSide implements AbstractBorderSide {
     return material.BorderSide(
       color: color!.build(),
       width: width,
+      style: style == null ? material.BorderStyle.solid : style!.build(),
     );
   }
 }